RL C Cartpole file update (#170)
* RL C Cartpole improvements (#171)
* Incorporating the updates made to the Ada code into the C cartpole environment.
* Removing the rest of the cartpole directory as we've merged the environment into the library.
