--- experiments/c/cartpole/Makefile
@@ -1,4 +1,0 @@
-.PHONY: build
-
-build: cartpole.c
-	gcc cartpole.c -g -Og -Wall -Wextra -lm -o cartpole
